Add pthread as an extra-libraries.
Without this, programs that have this library as a dependency need to
link pthread explicitly in their own cabal files. But that's arguably
a leak of internal implementation details.

What's weird is that downstream programs need to link pthread
explicitly only on some systems, not all. But this ought to fix it.