Export Internal module.
Following convention of many other Haskell packages out there, best to
export more rather than less. The internal modules are clearly named
as such so no API stability guarantees are made.

As a benefit the test-suite section gets to be simpler, shorter and
the compile times slightly faster, since we don't need to recompile
the whole library for tests.